requirements.txt
alembic==1.8.1
astroid==2.12.13
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.12.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==7.0.0
dill==0.3.6
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
filelock==3.8.2
flake8==6.0.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
freezegun==1.2.2
greenlet==2.0.1
gunicorn==20.0.4
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
isort==5.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
pip==22.3.1
platformdirs==2.6.0
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.15.9
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
pyproject_api==1.2.1
pytest==7.2.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
setuptools==65.5.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.11.0
